requirements.txt
Flaskmysql-connector-pythonpandasscikit-learnyfinance